cdc: use the new generation management service
Remove the old CDC management code from storage_service and instead use
the new CDC generation management service by simply calling two of its
methods in appropriate places.

The CDC log service (`cdc_service`) now obtains cdc::metadata from the
generation management service instead of storage_service.

Initialize the service and pass it to other services in the startup code
in main.cc and tests.

These refactors also change (simplify) the startup procedure slightly.
Previously, a new CDC generation would be always created on the
auto_bootstrap=off/seed joining path. However, we would not be able to
ensure that it contains the tokens of all nodes. Therefore this part was
removed. Now it will only create a generation on rolling upgrade or if
its the very first node.

If the user joins a node as a seed or with auto_bootstrap=off
into a cluster, the old generation will be used, which means colocation
of streams will be lost. To restore it, the user will need to use a
nodetool command to manually create a new CDC generation (separate PR).